+++ src/lesson2/task2/Logical.kt
@@ -30,7 +30,7 @@ fun isNumberHappy(number: Int): Boolean = (number % 10 + number / 10 % 10) == (n
  * Определить, угрожают ли они друг другу. Вернуть true, если угрожают.
  * Считать, что ферзи не могут загораживать друг друга.
  */
-fun queenThreatens(x1: Int, y1: Int, x2: Int, y2: Int): Boolean = ((x1 == x2) || (y1 == y2) || (abs(x1 - x2) / abs(y1 - y2) == 1))
+fun queenThreatens(x1: Int, y1: Int, x2: Int, y2: Int): Boolean = ((x1 == x2) || (y1 == y2) || ((abs(x1 - x2).toDouble() / abs(y1 - y2) == 1.0)))
 
 /**
  * Простая
